+++ test/context_test.py
@@ -1,0 +1,5 @@
+import os
+import sys
+# Se hace referencia al directorio 
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+import src
